requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
retina-face>=0.0.1
insightface>=0.7.0
faiss-cpu>=1.7.0
onnxruntime>=1.10.0